feat: algorithm

+++ crates/dies-world/src/filter/matrix_gen.rs
@@ -6,10 +6,10 @@ use na::allocator::Allocator;
 
 type U4 = Const<4>;
 fn block_diag(
-    a: &OMatrix<f32, U2, U2>,
-) -> OMatrix<f32, U4, U4>
+    a: &OMatrix<f64, U2, U2>,
+) -> OMatrix<f64, U4, U4>
 {
-    let mut result = OMatrix::<f32, U4, U4>::zeros();
+    let mut result = OMatrix::<f64, U4, U4>::zeros();
     result.fixed_slice_mut::<2, 2>(0, 0).copy_from(a);
     result.fixed_slice_mut::<2, 2>(2, 2).copy_from(a);
     result
@@ -18,9 +18,9 @@ fn block_diag(
 pub trait MatrixCreator<D>
     where
         D: DimName,
-        DefaultAllocator: Allocator<f32, D, D>,
+        DefaultAllocator: Allocator<f64, D, D>,
 {
-    fn create_matrix(&self, delta_t: f32) -> OMatrix<f32, D, D>;
+    fn create_matrix(&self, delta_t: f64) -> OMatrix<f64, D, D>;
 }
 
 struct Piecewise1stOrder;
@@ -28,8 +28,8 @@ struct Piecewise1stOrder;
 // 2x2 measurement noise matrix assuming acceleration is different in
 // different duration
 impl MatrixCreator<U2> for Piecewise1stOrder {
-    fn create_matrix(&self, delta_t: f32) -> OMatrix<f32, U2, U2> {
-        OMatrix::<f32, U2, U2>::new(
+    fn create_matrix(&self, delta_t: f64) -> OMatrix<f64, U2, U2> {
+        OMatrix::<f64, U2, U2>::new(
             delta_t.powi(4)/4.0, delta_t.powi(3)/2.0,
             delta_t.powi(3)/2.0, delta_t.powi(2),
         )
@@ -38,8 +38,8 @@ impl MatrixCreator<U2> for Piecewise1stOrder {
 
 //Piecewise1stOrder but 2D
 impl MatrixCreator<U4> for Piecewise1stOrder {
-    fn create_matrix(&self, delta_t: f32) -> OMatrix<f32,U4,U4> {
-        let m = OMatrix::<f32, U2, U2>::new(
+    fn create_matrix(&self, delta_t: f64) -> OMatrix<f64,U4,U4> {
+        let m = OMatrix::<f64, U2, U2>::new(
             delta_t.powi(4)/4.0, delta_t.powi(3)/2.0,
             delta_t.powi(3)/2.0, delta_t.powi(2),
         );
@@ -51,8 +51,8 @@ struct WhiteNoise1stOrder;
 
 // 2x2 measurement noise matrix assuming acceleration does not change
 impl MatrixCreator<U2> for WhiteNoise1stOrder {
-    fn create_matrix(&self, delta_t: f32) -> OMatrix<f32, U2, U2> {
-        OMatrix::<f32, U2, U2>::new(
+    fn create_matrix(&self, delta_t: f64) -> OMatrix<f64, U2, U2> {
+        OMatrix::<f64, U2, U2>::new(
             delta_t.powi(3)/3.0,  delta_t.powi(2)/2.0,
             delta_t.powi(2)/2.0, delta_t,
         )
@@ -61,8 +61,8 @@ impl MatrixCreator<U2> for WhiteNoise1stOrder {
 
 //WhiteNoise1stOrder but 2D
 impl MatrixCreator<U4> for WhiteNoise1stOrder {
-    fn create_matrix(&self, delta_t: f32) -> OMatrix<f32,U4,U4> {
-        let m = OMatrix::<f32, U2, U2>::new(
+    fn create_matrix(&self, delta_t: f64) -> OMatrix<f64,U4,U4> {
+        let m = OMatrix::<f64, U2, U2>::new(
             delta_t.powi(3)/3.0,  delta_t.powi(2)/2.0,
             delta_t.powi(2)/2.0, delta_t,
         );
@@ -74,8 +74,8 @@ struct ULMotionModel;
 
 // 2x2 transition matrix assuming constant speed
 impl MatrixCreator<U2> for ULMotionModel {
-    fn create_matrix(&self, delta_t: f32) -> OMatrix<f32, U2, U2> {
-        OMatrix::<f32, U2, U2>::new(
+    fn create_matrix(&self, delta_t: f64) -> OMatrix<f64, U2, U2> {
+        OMatrix::<f64, U2, U2>::new(
             1.0, delta_t,
             0.0, 1.0,
         )
@@ -84,8 +84,8 @@ impl MatrixCreator<U2> for ULMotionModel {
 
 //ULMotionModel but 2D
 impl MatrixCreator<U4> for ULMotionModel {
-    fn create_matrix(&self, delta_t: f32) -> OMatrix<f32,U4,U4> {
-        let m = OMatrix::<f32, U2, U2>::new(
+    fn create_matrix(&self, delta_t: f64) -> OMatrix<f64,U4,U4> {
+        let m = OMatrix::<f64, U2, U2>::new(
             1.0, delta_t,
             0.0, 1.0,
         );
@@ -96,8 +96,8 @@ impl MatrixCreator<U4> for ULMotionModel {
 
 
 // observation transformation matrix 1d [x, vx] -> [x]
-fn state_to_measurement() -> SMatrix<f32, 1, 2> {
-    SMatrix::<f32, 1, 2>::new(1.0, 0.0)
+fn state_to_measurement() -> SMatrix<f64, 1, 2> {
+    SMatrix::<f64, 1, 2>::new(1.0, 0.0)
 }
 
 #[cfg(test)]
@@ -106,9 +106,9 @@ mod tests {
     #[test]
     fn test_white_noise_1st_order_u2() {
         let wn = WhiteNoise1stOrder;
-        let delta_t = 2.0_f32;
-        let result_matrix:OMatrix<f32,U2,U2> = wn.create_matrix(delta_t);
-        let expected_matrix = OMatrix::<f32, U2, U2>::new(
+        let delta_t = 2.0_f64;
+        let result_matrix:OMatrix<f64,U2,U2> = wn.create_matrix(delta_t);
+        let expected_matrix = OMatrix::<f64, U2, U2>::new(
             8.0/3.0, 4.0/2.0,
             4.0/2.0, 2.0,
         );
@@ -119,9 +119,9 @@ mod tests {
     #[test]
     fn test_white_noise_1st_order_u4() {
         let wn = WhiteNoise1stOrder;
-        let delta_t = 1.0_f32;
-        let result_matrix:OMatrix<f32,U4,U4> = wn.create_matrix(delta_t);
-        let expected_matrix = OMatrix::<f32, U4, U4>::new(
+        let delta_t = 1.0_f64;
+        let result_matrix:OMatrix<f64,U4,U4> = wn.create_matrix(delta_t);
+        let expected_matrix = OMatrix::<f64, U4, U4>::new(
             1.0/3.0, 1.0/2.0, 0.0, 0.0,
             1.0/2.0, 1.0, 0.0, 0.0,
             0.0, 0.0, 1.0/3.0, 1.0/2.0,
